requirements.txt
redis
django-cors-headers
requests
pika
cassandra-driver
lz4
scales
mysqlclient
SpeechRecognition